backend/src/api/weather-tile-server/services/surf-api/requirements.txt
fastapi
uvicorn[standard]
mercantile
numpy
Pillow
cachetools
python-dotenv
google-cloud-datastore
gcsfs
zarr
